split to two file
